settings components folder reorganization
